refactor(api): move source-specific noise filtering to respective extractors
