+++ functions/api/utils.ts
@@ -12,22 +12,6 @@ export function decodeHtmlEntities(text: string): string {
     .replace(/&#x([0-9a-f]+);/gi, (_, hex) => String.fromCharCode(parseInt(hex, 16)));
 }
 
-export function isNoise(text: string): boolean {
-  const NOISE_KEYWORDS = [
-    "live coverage",
-    "Thank you for joining us",
-    "seen by the ABC",
-    "asked that the ABC use",
-    "Follow our live",
-    "Read more",
-    "More to come",
-    "Loading..."
-  ];
-  if (NOISE_KEYWORDS.some(kw => text.includes(kw))) return true;
-  if (text.length < 20) return true;
-  return false;
-}
-
 export function smartTruncate(text: string, maxLength: number): string {
   if (text.length <= maxLength) return text;
   const truncated = text.substring(0, maxLength);
